feat: implementar vista principal de maquinaria con filtros y tabla de datos
